requirements.txt
aiogram==3.0.0b7
aiohttp==3.8.3
python-dateutil==2.8.2